fix favorites
The display_history_and_favorites() function should only display the history and favorites items. It should not be responsible for deleting items or updating the session state. You can create two separate functions for those tasks, such as delete_history_item() and delete_favorite_item().
The st.button() method should be replaced with st.checkbox() to allow the user to add an item to favorites and also to remove it if needed. You can add a unique key to each checkbox to keep track of the state of the checkbox.
When the user adds an item to favorites, you should create a copy of the dictionary and append it to the favorites list. This is because dictionaries are mutable and if you append the same dictionary to the list multiple times, any changes to one of the dictionaries will be reflected in all of them.